+++ Changelog.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [0.16.3] - 2025-14-10
+
+### Added
+  - Configuration:
+    * Support "subdir" for target which can split their output in multiple file to ssave generated sub-file in a subdirectory.
+      The subdir can be provided at command-line level, at the top configuration file level or inside the target section of the configuration file
+    * Support a json section with gen_inc, local and subdir settings
+  - Fixed:
+    * JSON: escape double-quote in description
+
+## [0.16.2] - 2025-10-10
+
+### Fixed
+  - Python: incorrect import in top rifmux
+
 ## [0.16.1] - 2025-08-21
 
 ### Changed